+++ hahomematic/caches/persistent.py
@@ -120,18 +120,6 @@ class DeviceDescriptionCache(BasePersistentCache):
         # {interface_id, {address, device_descriptions}}
         self._device_descriptions: Final[dict[str, dict[str, dict[str, Any]]]] = {}
 
-    def _add_device_descriptions(
-        self, interface_id: str, device_descriptions: list[dict[str, Any]]
-    ) -> None:
-        """Add device_descriptions to cache."""
-        if interface_id not in self._raw_device_descriptions:
-            self._raw_device_descriptions[interface_id] = []
-        self._raw_device_descriptions[interface_id] = device_descriptions
-
-        self._convert_device_descriptions(
-            interface_id=interface_id, device_descriptions=device_descriptions
-        )
-
     def add_device_description(
         self, interface_id: str, device_description: dict[str, Any]
     ) -> None:
@@ -139,8 +127,10 @@ class DeviceDescriptionCache(BasePersistentCache):
         if interface_id not in self._raw_device_descriptions:
             self._raw_device_descriptions[interface_id] = []
 
-        if device_description not in self._raw_device_descriptions[interface_id]:
-            self._raw_device_descriptions[interface_id].append(device_description)
+        self._remove_device(
+            interface_id=interface_id, deleted_addresses=[device_description[HM_ADDRESS]]
+        )
+        self._raw_device_descriptions[interface_id].append(device_description)
 
         self._convert_device_description(
             interface_id=interface_id, device_description=device_description
@@ -154,26 +144,25 @@ class DeviceDescriptionCache(BasePersistentCache):
         """Remove device from cache."""
         deleted_addresses: list[str] = [device.device_address]
         deleted_addresses.extend(device.channels)
-        self._add_device_descriptions(
-            interface_id=device.interface_id,
-            device_descriptions=[
-                raw_device
-                for raw_device in self.get_raw_device_descriptions(device.interface_id)
-                if raw_device[HM_ADDRESS] not in deleted_addresses
-            ],
-        )
+        self._remove_device(interface_id=device.interface_id, deleted_addresses=deleted_addresses)
+        await self.save()
+
+    def _remove_device(self, interface_id: str, deleted_addresses: list[str]) -> None:
+        """Remove device from cache."""
+        self._raw_device_descriptions[interface_id] = [
+            raw_device
+            for raw_device in self.get_raw_device_descriptions(interface_id)
+            if raw_device[HM_ADDRESS] not in deleted_addresses
+        ]
 
         for address in deleted_addresses:
             try:
-                if ":" not in address and self._addresses.get(device.interface_id, {}).get(
-                    address, []
-                ):
-                    del self._addresses[device.interface_id][address]
-                if self._device_descriptions.get(device.interface_id, {}).get(address, {}):
-                    del self._device_descriptions[device.interface_id][address]
+                if ":" not in address and self._addresses.get(interface_id, {}).get(address, []):
+                    del self._addresses[interface_id][address]
+                if self._device_descriptions.get(interface_id, {}).get(address, {}):
+                    del self._device_descriptions[interface_id][address]
             except KeyError:
                 _LOGGER.warning("REMOVE_DEVICE failed: Unable to delete: %s", address)
-        await self.save()
 
     def get_addresses(self, interface_id: str) -> dict[str, list[str]]:
         """Return the addresses by interface."""