+++ hahomematic/platforms/device.py
@@ -13,7 +13,10 @@ from hahomematic.const import (
     EVENT_CONFIG_PENDING,
     EVENT_STICKY_UN_REACH,
     EVENT_UN_REACH,
+    HM_AVAILABLE_FIRMWARE,
     HM_FIRMWARE,
+    HM_FIRMWARE_UPDATABLE,
+    HM_FIRMWARE_UPDATE_STATE,
     HM_SUBTYPE,
     HM_TYPE,
     INIT_DATETIME,
@@ -23,6 +26,7 @@ from hahomematic.const import (
     PARAMSET_KEY_VALUES,
     RELEVANT_INIT_PARAMETERS,
     HmCallSource,
+    HmDeviceFirmwareState,
     HmForcedDeviceAvailability,
 )
 from hahomematic.exceptions import BaseHomematicException
@@ -37,6 +41,7 @@ from hahomematic.platforms.support import (
     get_device_name,
     value_property,
 )
+from hahomematic.platforms.update import HmUpdate
 from hahomematic.support import updated_within_seconds
 
 _LOGGER = logging.getLogger(__name__)
@@ -68,6 +73,7 @@ class HmDevice(PayloadMixin):
         self._attr_last_update: datetime = INIT_DATETIME
         self._forced_availability: HmForcedDeviceAvailability = HmForcedDeviceAvailability.NOT_SET
         self._update_callbacks: Final[list[Callable]] = []
+        self._firmware_update_callbacks: Final[list[Callable]] = []
         self._attr_device_type: Final = str(
             self.central.device_descriptions.get_device_parameter(
                 interface_id=interface_id,
@@ -86,13 +92,6 @@ class HmDevice(PayloadMixin):
         self._has_custom_entity_definition: Final = cep.has_custom_entity_definition_by_device(
             device=self
         )
-        self._attr_firmware: Final = str(
-            self.central.device_descriptions.get_device_parameter(
-                interface_id=interface_id,
-                device_address=device_address,
-                parameter=HM_FIRMWARE,
-            )
-        )
 
         self._attr_name: Final = get_device_name(
             central=central,
@@ -101,13 +100,53 @@ class HmDevice(PayloadMixin):
         )
         self.value_cache: Final = ValueCache(device=self)
         self._attr_room: Final = central.device_details.get_room(device_address=device_address)
-
+        self._update_firmware_data()
+        self._attr_update_entity: Final = HmUpdate(device=self)
         _LOGGER.debug(
             "__INIT__: Initialized device: %s, %s, %s, %s",
             self._attr_interface_id,
             self._attr_device_address,
             self._attr_device_type,
             self._attr_name,
+        )
+
+    def _update_firmware_data(self) -> None:
+        """Update firmware related data from device descriptions."""
+        self._attr_available_firmware: str | None = (
+            self.central.device_descriptions.get_device_parameter(
+                interface_id=self._attr_interface_id,
+                device_address=self._attr_device_address,
+                parameter=HM_AVAILABLE_FIRMWARE,
+            )
+            or None
+        )
+        self._attr_firmware = str(
+            self.central.device_descriptions.get_device_parameter(
+                interface_id=self._attr_interface_id,
+                device_address=self._attr_device_address,
+                parameter=HM_FIRMWARE,
+            )
+        )
+
+        try:
+            self._attr_firmware_update_state = HmDeviceFirmwareState(
+                str(
+                    self.central.device_descriptions.get_device_parameter(
+                        interface_id=self._attr_interface_id,
+                        device_address=self._attr_device_address,
+                        parameter=HM_FIRMWARE_UPDATE_STATE,
+                    )
+                )
+            )
+        except ValueError:
+            self._attr_firmware_update_state = HmDeviceFirmwareState.UP_TO_DATE
+
+        self._attr_firmware_updatable = bool(
+            self.central.device_descriptions.get_device_parameter(
+                interface_id=self._attr_interface_id,
+                device_address=self._attr_device_address,
+                parameter=HM_FIRMWARE_UPDATABLE,
+            )
         )
 
     @value_property
@@ -121,6 +160,11 @@ class HmDevice(PayloadMixin):
         if un_reach is not None and un_reach.value is not None:
             return not un_reach.value
         return True
+
+    @config_property
+    def available_firmware(self) -> str | None:
+        """Return the available firmware of the device."""
+        return self._attr_available_firmware
 
     @property
     def config_pending(self) -> bool:
@@ -143,6 +187,16 @@ class HmDevice(PayloadMixin):
     def firmware(self) -> str:
         """Return the firmware of the device."""
         return self._attr_firmware
+
+    @config_property
+    def firmware_updatable(self) -> bool:
+        """Return the firmware update state of the device."""
+        return self._attr_firmware_updatable
+
+    @config_property
+    def firmware_update_state(self) -> HmDeviceFirmwareState:
+        """Return the firmware update state of the device."""
+        return self._attr_firmware_update_state
 
     @config_property
     def interface(self) -> str:
@@ -173,6 +227,11 @@ class HmDevice(PayloadMixin):
     def sub_type(self) -> str:
         """Return the sub_type of the device."""
         return self._attr_sub_type
+
+    @property
+    def update_entity(self) -> HmUpdate:
+        """Return the device firmware update entity of the device."""
+        return self._attr_update_entity
 
     @property
     def _e_unreach(self) -> GenericEntity | None:
@@ -250,6 +309,19 @@ class HmDevice(PayloadMixin):
         if update_callback in self._update_callbacks:
             self._update_callbacks.remove(update_callback)
 
+    def register_firmware_update_callback(self, firmware_update_callback: Callable) -> None:
+        """Register firmware update callback."""
+        if (
+            callable(firmware_update_callback)
+            and firmware_update_callback not in self._firmware_update_callbacks
+        ):
+            self._firmware_update_callbacks.append(firmware_update_callback)
+
+    def unregister_firmware_update_callback(self, firmware_update_callback: Callable) -> None:
+        """Remove firmware update callback."""
+        if firmware_update_callback in self._firmware_update_callbacks:
+            self._firmware_update_callbacks.remove(firmware_update_callback)
+
     def _set_last_update(self) -> None:
         self._attr_last_update = datetime.now()
 
@@ -279,6 +351,28 @@ class HmDevice(PayloadMixin):
             interface_id=self._attr_interface_id,
             device_address=self._attr_device_address,
         )
+
+    def refresh_firmware_data(self) -> None:
+        """Refresh firmware data of the device."""
+        old_available_firmware = self._attr_available_firmware
+        old_firmware = self._attr_firmware
+        old_firmware_update_state = self._attr_firmware_update_state
+        old_firmware_updatable = self._attr_firmware_updatable
+
+        self._update_firmware_data()
+
+        if (
+            old_available_firmware != self._attr_available_firmware
+            or old_firmware != self._attr_firmware
+            or old_firmware_update_state != self._attr_firmware_update_state
+            or old_firmware_updatable != self._attr_firmware_updatable
+        ):
+            for _firmware_callback in self._firmware_update_callbacks:
+                _firmware_callback()
+
+    async def update_firmware(self) -> bool:
+        """Update the firmware of the homematic device."""
+        return await self.client.update_device_firmware(device_address=self._attr_device_address)
 
     async def load_value_cache(self) -> None:
         """Init the parameter cache."""