+++ changelog.md
@@ -1,3 +1,8 @@
+# Version 2023.5.1 (2023-05-11)
+
+- Refactor device description cache handling
+- Add device firmware update handling
+
 # Version 2023.5.0 (2023-05-01)
 
 - Remove unsupported on_time / refactor HmIP-RGBW